+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 2.3.3
 
 - add security header also when responding with a redirect
+- updated the way to disable http2
 
 ## 2.3.2